+++ src/main/java/sh2/dict/Sh2Dict.java
@@ -126,7 +126,7 @@ public class Sh2Dict {
         public static final int REG_MASK = SH2_REG_MASK;
 
         RegSpec(int addr, String name, Size size) {
-            this(addr, name, size, (int) size.getMask());
+            this(addr, name, size, size.getMask());
         }
 
         RegSpec(int addr, String name, Size size, int writeMask) {
